requirements.txt
flask==2.3.3
flask-cors==4.0.0
firebase-admin==6.2.0
python-dotenv==1.0.0
werkzeug==2.3.7
google-generativeai==0.3.1
requests==2.28.2
pytz==2024.1
gunicorn==21.2.0
flask_caching==2.0.0